Change the algorithm for where to find package.son in serviceDetails
This commit fixes a problem in Windows.

When hosting on Windows on IIS in iisnode, process.mainModule isn't the
application. Instead iisnode's interceptor.js is started as main, which
then loads the application. This means process.mainModule.paths cannot
be used.

If we're on iisnode we can use the fact that current working directory
is set to the application's and we look for package.json there.

Unfortunately, there is no safe way to determine if we're on iisnode.
To make sure msb properly can detect that we're using iisnode the
env var IISNODE_HOSTED can be set to any value. (Or really, any env var
that starts with "IISNODE_" can be set)